Move NestedTraversalFinder to its own header
